Доработка репозитория Category
