+++ app/src/main/java/com/mashjulal/android/financetracker/domain/financialcalculations/Category.kt
@@ -2,4 +2,6 @@ package com.mashjulal.android.financetracker.domain.financialcalculations
 
 import android.support.annotation.DrawableRes
 
-data class Category(var operationType: OperationType, val title: String, @DrawableRes val imageRes: Int)
+data class Category(var operationType: OperationType = OperationType.OUTGOINGS,
+                    val title: String = "Dummy Category",
+                    @DrawableRes val imageRes: Int = 0)
