+++ app/src/main/java/com/mashjulal/android/financetracker/domain/repository/sqliteimpl/CategoryRepositoryImpl.kt
@@ -1,11 +1,10 @@
 package com.mashjulal.android.financetracker.domain.repository.sqliteimpl
 
-import android.support.annotation.DrawableRes
-import com.mashjulal.android.financetracker.R
 import com.mashjulal.android.financetracker.domain.financialcalculations.Category
 import com.mashjulal.android.financetracker.domain.financialcalculations.OperationType
 import com.mashjulal.android.financetracker.domain.repository.CategoryRepository
 import com.mashjulal.android.financetracker.domain.repository.sqliteimpl.model.InnerCategory
+import com.mashjulal.android.financetracker.domain.repository.sqliteimpl.utils.CategoryMapper
 import com.mashjulal.android.financetracker.domain.repository.sqlmodel.CategoryModel
 import com.squareup.sqldelight.SqlDelightQuery
 import io.reactivex.Completable
@@ -15,38 +14,7 @@ import javax.inject.Inject
 
 class CategoryRepositoryImpl @Inject constructor(val core: SQLiteCore) : CategoryRepository {
 
-    companion object {
 
-        const val PREDEFINED_OTHER_INCOMINGS = "OtherIncomings"
-        const val PREDEFINED_OTHER_OUTGOINGS = "OtherOutgoings"
-        const val PREDEFINED_FROM_ACCOUNT = "FromAccount"
-        const val PREDEFINED_TO_ACCOUNT = "ToAccount"
-
-
-        val mapResourceIDtoString = hashMapOf(
-                R.drawable.ic_cash_back_green_24dp to PREDEFINED_OTHER_INCOMINGS,
-                R.drawable.ic_salary_green_24dp to "Salary",
-                R.drawable.ic_card_giftcard_green_24dp to "Gift",
-                R.drawable.ic_bills_red_24dp to PREDEFINED_OTHER_OUTGOINGS,
-                R.drawable.ic_videogame_red_24dp to "Entertainments",
-                R.drawable.ic_car_red_24dp to "Transport",
-                R.drawable.ic_shopping_red_24dp to "Shopping"
-        )
-
-        val mapStringToResourceID = hashMapOf(
-                "OtherIncomings" to R.drawable.ic_cash_back_green_24dp,
-                "Salary" to R.drawable.ic_salary_green_24dp,
-                "Gift" to R.drawable.ic_card_giftcard_green_24dp,
-                "OtherOutgoings" to R.drawable.ic_bills_red_24dp,
-                "Entertainments" to R.drawable.ic_videogame_red_24dp,
-                "Transport" to R.drawable.ic_car_red_24dp,
-                "Shopping" to R.drawable.ic_shopping_red_24dp
-        )
-
-        fun getStringByRId(@DrawableRes key: Int) = mapResourceIDtoString[key] ?: "OtherOutgoings"
-        fun getRIdByString(@DrawableRes key: String) = mapStringToResourceID[key]
-                ?: R.drawable.ic_bills_red_24dp
-    }
 
 
     override fun getAll(): Single<List<Category>> {
@@ -54,10 +22,7 @@ class CategoryRepositoryImpl @Inject constructor(val core: SQLiteCore) : Categor
         return core.database.createQuery(CategoryModel.TABLE_NAME, statement.sql)
                 .mapToList { it -> InnerCategory.ALL_CATEGORIES_MAPPER.map(it) }
                 .map { it ->
-                    it.map { it ->
-                        Category(OperationType.getTypeByString(it.type()), it.category(),
-                                getRIdByString(it.subcategory()))
-                    }.toList()
+                    it.map { it -> CategoryMapper.newCategory(it) }.toList()
                 }.take(1).single(listOf())
     }
 
@@ -66,8 +31,7 @@ class CategoryRepositoryImpl @Inject constructor(val core: SQLiteCore) : Categor
         return core.database.createQuery(CategoryModel.TABLE_NAME, statement.sql)
                 .mapToOne { it -> InnerCategory.SELECT_CATEGORY_BY_TYPE.map(it) }
                 .map { it ->
-                    Category(OperationType.getTypeByString(it.type()), it.category(),
-                            getRIdByString(it.subcategory()))
+                    CategoryMapper.newCategory(it)
                 }
                 .collect({ mutableListOf() }, { b: MutableList<Category>, t: Category -> b.add(t) })
                 .map { it.toList() }
@@ -78,7 +42,7 @@ class CategoryRepositoryImpl @Inject constructor(val core: SQLiteCore) : Categor
             val db = core.database.writableDatabase
             val insertCategory = CategoryModel.InsertCategory(db)
             insertCategory.bind(category.title, category.operationType.toString(),
-                    getStringByRId(category.imageRes))
+                    CategoryMapper.getStringByRId(category.imageRes))
             insertCategory.executeInsert()
         }
     }
