+++ build.gradle
@@ -26,7 +26,12 @@ allprojects {
     repositories {
         google()
         jcenter()
-        maven { url 'https://jitpack.io' }
+        maven {
+            url "https://plugins.gradle.org/m2/"
+        }
+        repositories {
+            maven { url 'https://jitpack.io' }
+        }
     }
 }
